requirements.txt
streamlit
langchain
python-dotenv
PyPDF2
faiss-cpu
openai
tiktoken
pickle5
